requirements.txt
requests==2.32.3
beautifulsoup4==4.13.4
selenium==4.31.0
python-dotenv==1.1.0